requirements.txt
asgiref==3.7.2
astroid==2.15.6
asttokens==2.3.0
async-timeout==4.0.3
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coverage==5.4
decorator==5.1.1
dill==0.3.7
distlib==0.3.7
Django==4.1.13
django-oscar==3.2.2
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-extra-views==0.14.0
django-haystack==3.2.1
django-phonenumber-field==6.4.0
django-redis==5.2.0
django-tables2==2.3.4
django-treebeard==4.7
django-webtest==1.9.10
django-widget-tweaks==1.5.0
django-import-export==4.1.1
easy-thumbnails==2.8.5
exceptiongroup==1.1.3
execnet==2.0.2
executing==1.2.0
factory-boy==3.2.1
Faker==19.3.1
filelock==3.12.3
freezegun==1.2.2
idna==3.4
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.15.0
isort==5.12.0
jedi==0.19.0
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
phonenumbers==8.13.19
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
purl==1.6
py==1.11.0
pycountry==22.3.5
Pygments==2.16.1
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyprof2calltree==1.4.5
pysolr==3.9.0
pytest==7.4.1
pytest-django==3.8.0
pytest-forked==1.6.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
sorl-thumbnail==12.9.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.12.1
tox==3.28.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.4
waitress==2.1.2
wcwidth==0.2.6
WebOb==1.8.7
WebTest==2.0.35
whitenoise==6.4.0
Whoosh==2.7.4
wrapt==1.15.0
gunicorn==21.2.0
djangorestframework==3.14.0
django-tinymce==3.6.1
django-cors-headers==4.3.1
django-oscar-api==3.2.4
django-oscar-paypal@git+https://github.com/django-oscar/django-oscar-paypal.git@04b77bdf1bdfb5a7123166a7ddeade6b08f0704b
stripe==9.11.0
qrcode==7.4.2
drf-spectacular==0.27.2
cryptography==43.0.1
firebase-admin==6.5.0
channels[daphne]==4.0.0
channels-redis==4.0.0